Feature/spike handling

+++ src/cell_tree.hpp
@@ -32,7 +32,7 @@ class cell_tree {
 public :
     // use a signed 16-bit integer for storage of indexes, which is reasonable given
     // that typical cells have at most 1000-2000 segments
-    using int_type        = int16_t;
+    using int_type        = int;
     using index_type      = memory::HostVector<int_type>;
     using view_type       = index_type::view_type;
     using const_view_type = index_type::const_view_type;
@@ -298,4 +298,3 @@ private :
 
 } // namespace mc
 } // namespace nest
-
